Fix 1 - Standardize data in classes.js
Standardizing the objects in classes.js fixes the missing image bug. It seems like the bug is in the “API” that’s providing the data not the code. Alternatively if the “API” is not buggy then content.featureImage should be set by testing for different properties in the “API” and using whichever is present to set the content.featureImage property before it’s used in the presentation.